Maze&LavaObstacles generation & Some experiments

+++ src/main/java/xyz/morecraft/dev/malmo/util/IntPoint3D.java
@@ -2,9 +2,9 @@ package xyz.morecraft.dev.malmo.util;
 
 public class IntPoint3D {
 
-    public int x, y, z;
+    public float x, y, z;
 
-    public IntPoint3D(int x, int y, int z) {
+    public IntPoint3D(float x, float y, float z) {
         this.x = x;
         this.y = y;
         this.z = z;
